Rename `touchlink` back into `lightlink`
